# swagger.yaml
openapi: 3.0.3
info:
  title: Pipu Pro
  description: >-
    # Pipu Pro  api documentation
  version: 1.0.0
  contact: { email: ammuftau74@gmail.com }
servers:
  - url: http://localhost:9000/api/v1
    description: This is my local Server
  - url: https://pipu-pro.onrender.com/api/v1
    description: This is the original api URL
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
paths:
...
  /messages:
    post:
      tags:
        - 🔄 General
      summary: send message
      description: >-
        The API endpoint allows you to send chat message.
      operationId: sendMessage
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                chatId:
                  type: string
                  example: 683c89c87f16c6acb1ca3d84
                content:
                  type: string
                  example: Hello, how are you?
      responses:
        '201':
          description: New message sent
          headers:
            Access-Control-Allow-Credentials:
              schema:
                type: string
                example: 'true'
            Access-Control-Allow-Origin:
              schema:
                type: string
                example: '*'
            Connection:
              schema:
                type: string
                example: keep-alive
            Content-Length:
              schema:
                type: string
                example: '500'
            Date:
              schema:
                type: string
                example: Fri, 30 Jun 2023 06:44:09 GMT
            ETag:
              schema:
                type: string
                example: W/"1f4-1q38Sbu25p0UVmeYQmrEv2VrgPg"
            Keep-Alive:
              schema:
                type: string
                example: timeout=5
            RateLimit-Limit:
              schema:
                type: string
                example: '500'
            RateLimit-Remaining:
              schema:
                type: string
                example: '495'
            RateLimit-Reset:
              schema:
                type: string
                example: '772'
            X-Powered-By:
              schema:
                type: string
                example: Express
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                  message:
                    type: string
                    example: >-
                      Message sent successfully.
...